add full skeleton for stage2

+++ boot/loader/include/vgafont.h
@@ -5,10 +5,10 @@
 #pragma once
 
 
-void DrawCharacter(char c, int x, int y, int fgcolor);
+void DrawCharacter(char c, INT x, INT y, WORD fgcolor);
 
-void DrawCharacterBackDrop(char c, int x, int y, int fgcolor, int bgcolor);
+void DrawCharacterBackDrop(char c, INT x, INT y, WORD fgcolor, WORD bgcolor);
 
-void DrawString(char* str, int x, int y, int fgcolor);
+void DrawString(char* str, INT x, INT y, WORD fgcolor);
 
-void DrawStringBackDrop(char* str, int x, int y, int fgcolor, int bgcolor);
+void DrawStringBackDrop(char* str, INT x, INT y, WORD fgcolor, WORD bgcolor);
